feat(home): drive problem list from DB instead of mock data
app/page.tsx를 Server Component로 전환해 problems 테이블에서 직접
필터·정렬·페이지네이션된 슬라이스를 가져온다. 클라이언트 부분은
ChallengesView로 분리해 URL 파라미터 갱신만 담당.

- lib/queries/problems.ts — fetchProblemsForList + 파라미터 파서
- arrayOverlaps로 태그 OR 필터, exists 서브쿼리로 status(solved/unsolved)
- 정답률은 1/average_tries 근사, NULL 안전
- 인증 사용자는 user_solved_problems 조인으로 done 표시 가능
- mock-problems.ts 제거 (참조 없음)

상태 필터의 'tried'는 schema에 미존재라 무시 (TODO).
